Re-enable NaCl toolchains when requested
With some minor fixes, the build is able to run and produce an irt_core
binary when configured with the Saigo toolchain.
Added documentation how to do this to the README.

Still keeping the 'if nacltools' guards I added so that sel_ldr can be
built without acquiring any NaCl toolchains.