Don't use --sysroot flag by default
For Linux toolchains (except x86-64 where I previously commented it
out), there were hard-coded sysroot flags referring to the paths of
Chromium's bundled dependencies. Get rid of this. On my machine the
cross toolchains work fine without it. But I added a
a 'sysroot' command line argument, so if anyone needs it they can
pass sysroot=... on the scons command line to configure a --sysroot flag.