NaClAddrSpaceFree: do nothing if memory prereserved
NaClAddrSpaceFree is supposed to unmap the whole memory range used in
the sandbox, but when the bootstrap helper pre-reserved it starting at
0, it may have skipped the first few pages that the kernel does not
allow to be mapped and we won't know how many pages were skipped. This
function is only used in tests so just skip the unmap to avoid breaking
the tests. Fixes run_trusted_mmap_test on ARM.